New api

+++ src/main/java/sfa/order_service/dto/response/BeetReportResponse.java
@@ -10,6 +10,6 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class BeetReportResponse {
     Double totalSales;
-    Integer totalOrder;
+    Long totalOrder;
     BeetRespForOrderDto beetRespForOrderDto;
 }
